STY: Move functions within _xobj_to_image to a private module
